moved all the reservation info from the reservation component into the Navbar.js component and removed the reservation component to correct the issues with changing the modal show state properly
